// src/utils/types.ts
export interface Instance {
    ZUID?: string;
    blueprintID?: string;
    blueprintZUID?: string;
    cancelledReason?: string;
    createdAt?: string;
    createdByUserZUID?: string;
    deletedAt?: string;
    domain?: string;
    ecoID?: string;
    ecoZUID?: string;
    id?: string;
    legacy?: string;
    name?: string;
    planID?: string;
    prefs?: string;
    randomHashID?: string;
    requiresTwoFactor?: string;
    screenshotURL?: string;
    updatedAt?: string;
    useTLS?: string;
}